+++ tools/cmtupd/CHANGELOG.md
@@ -10,6 +10,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com)
 and this project adheres to [Semantic Versioning](http://semver.org).
 
+## cmtupd 1.1.2 (2024-07-31)
+- Fixed - Race printing final report
+- Added - Include files unchanged in final report
+
 ## cmtupd 1.1.1 (2024-04-19)
 - Fixed - Crash showing complex configuration